Enable libwebsockets compilation for *NIX
CMake is invoked in a somewhat hacky manner. It works but a better solution would be appreciated. The basic idea is to use CMake to configure libwebsockets in configure.ac, and then add the result as a subdirectory in Makefile.am. Autotools also forces the "build" subdirectory to pre-exist in order to work. Include an empty .gitignore file to force the subdirectory to be included.

In addition, compile Server.cpp so that ArmoryDB will link properly.