BIP 151 fixups / test additions
- Add more tests.
- Fix some underlying bugs.
- Miscellaneous cleanup.
- New test data based on an enhanced Bcoin test suite run.

Note that the test vectors deviate significantly from Bcoin. In particular, there seem to be some fundamental issues with somebody's library. Rekeys produce the same K1/K2 keys, and the first packet encryption/decryption matches (other than the Poly1305 tag). Otherwise, something deep is buried in one or both of the crypto libraries that isn't correct. (Bcrypto, the crypto library for Bcoin, appears to try to support the OpenSSH and RFC 7539 variants. This may be the point of failure.) For now, Armory can encrypt and decrypt its own packets. More research can be done later if needed.