rename config Armory -> ArmoryDB + minor changes
Rename the config directory `Armory` to `ArmoryDB`, add the `mkdir_p()`
utility function and use it to create the config directory.

Change the default config type from FULL to SUPERNODE.

In `processArgs()` call `startupBIP150CTX(4, false);` if the arg
`fullbip150` is present.

In `main()` change `startupBIP150CTX(4, false);` to `true`.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>