cmake support for Visual Studio/vcpkg
Add support for building ArmoryDB on Windows with Visual Studio using
cmake.

Dependencies are handled with vcpkg, and the `VCPKG_TARGET_TRIPLET`
variable can activate vcpkg support on other platforms as well.

Also fix the logic for the handling of string_option variables, this was
incorrect in the previous iteration of the cmake code.

Also remove the `-D` prefix for defines passed to
`target_compile_definitions()`, as this is unnecessary, and MSVC uses a
different syntax.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>