add cmake install rules
Add full support for `make install` as in the original autotools code
for armory.

Include the FindPython2 cmake modules as they are not in the base cmake
package on Ubuntu.

Add a couple of missing error checks to the cmake code.

Make some minor fixes to the autotools code as well, for things broken
by adding the cmake support and some other fixes.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>