cmake: bundle libwebsockets
Currently the Ubuntu version of libwebsockets is older than we need and
vcpkg fails to build it on linux.

Include a dump of the current libwebsockets master as a subdirectory and
add it via `add_subdirectory()` in cmake, which it fully supports.

To later check if we are including a bundled copy, we check:

```
IF(TARGET websockets)
```

and fall back to vcpkg or pkg-config versions. This allows simple
removal of this sub-tree in the future.

All libwebsockets cmake configuration options are passed through to its
cmake code.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>