fix some compatibility issues
Make cmake code work with cmake versions down to 2.8.12 which is on
Ubuntu 14 (trusty).

Add `-std=c99` to compiler options for libbtc for gcc and clang,
required on Ubuntu 14.

Include `sys/types.h` in libbtc for `ssize_t` if available, required on
Ubuntu 14.

Unfortunately Ubuntu 14 cannot be supported because the version of
protocol buffers is too old, 16 will be the lowest supported version.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>