limit max thread count to 126
When the number of threads is higher than 126 the number of LMDB locks
is exhausted and `ArmoryDB` dies with the error:

```
MDB_READERS_FULL: Environment maxreaders limit reached
```

See:
https://github.com/lmdbjava/lmdbjava/issues/65#issuecomment-322761020

Replace the uses of `std::thread::hardware_concurrency()` with a new
function `MAX_THREADS()` which limits the number of threads to `126`.

This is a temporary fix, the real fix would be making sure the number of
LMDB locks is sufficient.

However, a configured value for the maximum number of threads would also
be useful.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>