add LTO and gcc/clang toolchain support
Add support for LTO (link-time-optimizations) for gcc, clang and msvc.

The compile and link options are added in `LTO.cmake`.

They are enabled with the `ENABLE_LTO` option which defaults to `OFF`
for `Debug` mode and `ON` otherwise.

Add the option to the table in `README.md`.

This also requires two helper modules to set the appropriate toolchain,
`GCCToolchain.cmake` for gcc and `LLVMToolchain.cmake` for clang. These
use heuristics to find the appropriate toolchain binaries.

It is safe to include these without checking the compiler type, they do
this themselves.

In the case of gcc, the toolchain is not necessary when not compiling
with LTO, in the case of clang the toolchain may be necessary for
non-LTO builds as well.

Include these files in both the main `CMakeLists.txt` and in `LTO.cmake`
as well.

Also default to the swig/python stuff (`-DWITH_CLIENT`) being `OFF` for
clang because it looks like swig sometimes fails with clang, and we
don't need this.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>